fix: clean up buildFederatedSchema
